+++ src/hu/emanuel/jeremi/antitower/effect/Snow.java
@@ -72,7 +72,7 @@ public class Snow extends Weather {
      * @param delta
      */
     @Override
-    public final void update(long delta) {
+    public final void update(double delta) {
         int speed = (int) (this.speed * delta);
         for (int i = 0; i < amount; i++) {
             // first layer
